update to vs2017

+++ src/Core/SPK_Group.cpp
@@ -270,7 +270,7 @@ namespace SPK
 		}
 	}
 
-	void Group::reallocate(size_t capacity)
+	void Group::reallocate(unsigned int capacity)
 	{
 		SPK_ASSERT(capacity != 0,"Group::reallocate(size_t) - Group capacity must not be 0");
